Add files via upload
✨ Excited to Share! ✨

I’ve just completed my first assignment in Next.js by building a simple "Hello World" application and running it on localhost:3000. 🎉

This milestone wouldn’t have been possible without the incredible guidance and support of Sir Amen Aalam, Sir Asharib Ali, Sir Nassem Shah and Ahmed Raza Sheikh 🙌

Thank you for helping me take my first step into the world of modern web development!



This is a simple Next.js component that displays the text HELLO WORLD! on the webpage. When this code is placed in the pages/index.js file, it will render this message when you access the root URL localhost:3000/

Here's a brief explanation:

export default function Home() exports the Home component as the default export, so it can be imported in other files.
The return` statement contains the JSX that will be rendered to the page. In this case, it's an <h1> element displaying HELLO WORLD!.

#NextJS #WebDevelopment #LearningJourney #Gratitude #TechSkills